create controller to refactor the code
